+++ src/Makefile.test.include
@@ -34,6 +34,7 @@ RAW_TEST_FILES = test/data/alertTests.raw
 GENERATED_TEST_FILES = $(JSON_TEST_FILES:.json=.json.h) $(RAW_TEST_FILES:.raw=.raw.h)
 
 BITCOIN_TESTS =\
+  test/arith_uint256_tests.cpp \
   test/bignum.h \
   test/alert_tests.cpp \
   test/allocator_tests.cpp \
@@ -67,7 +68,6 @@ BITCOIN_TESTS =\
   test/test_bitcoin.cpp \
   test/timedata_tests.cpp \
   test/transaction_tests.cpp \
-  test/uint256_tests.cpp \
   test/univalue_tests.cpp \
   test/util_tests.cpp
